requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
numpy==2.3.3
openai==1.109.1
pandas==2.3.2
psycopg2==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
ruff==0.13.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
